requirements.txt
black
flake8
ipython
isort
jupyterlab
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
mlflow
dvclive
